Add `canonifyurls` config option.
Be able to inhibit AbsURL canonicalization of content, on a site
configuration basis.  Advantages of being able to inhibit this include
making it easier to rendering on other hostnames, and being able to
include resources on http or https depending on how this page was
retrieved, avoiding mixed-mode client complaints without adding latency
for plain http.