No markdown for Pygments
Treating the content of the highlight-shortcode as Markdown has created some issues; one had to indent the source code with 4 spaces to make it look OK.

There is a ongoing debate in the issues below about changing the shortcode API to mark shortcodes that does not want/need Markdown processing.

This commit skips that whole discussion by fixing just the code higlighter shortcode, by adding a new internal Set for shortcodes that should not be processed as Markdown.

Relates to #560 and #565