Removed the timezone offset from buildDate
The ‘%z’ adds the timezone offset to the buildDate
But this is not parsed by version.go and the result is and output like:

Hugo Static Site Generator v0.13-DEV-7831690 buildDate: 0001-01-01T00:00:00Z

Which is of course not what I wanted.

The reason for adding the timezone was, that buiding without the added llinker flags
formats the date like this:

Hugo Static Site Generator v0.13-DEV buildDate: 2014-11-25T22:43:53+07:00

So for some reason the date format is different when using the file modification time.

However, in my oppinion the important information for tracing source of the binary
is the git commit sha. Thus, I guess it is acceptable to use the local time for the build date.

Another option would be to use UTC for the build date and convert the output to local time at runtime in the version command.