Convert pages on modified templates and data files
We don't need to re-read the files from disk when shortcodes or data
files changes in server-mode, but we need to run the convert step.

To make this work, a copy is made of the original content when running
in watch mode, to use as a template so to speak.

Fixes #1971