Optimized main hugo.png image
A combination of lossless and lossy optimization reduces the image size from 49,506 bytes to 18,210 bytes (courtesy of TinyPNG.com and FileOptimizer). The nature of the image, being mostly a solid graphic with same or similar colors, lends itself to "lossy" compression that is visually lossless – I can't tell any difference between the 49 KB and 18 KB versions.