Add example for taxonomy terms with metadata
There is no information on how you can add custom metadata to your
taxonomy terms. What you need to do is add front matter data in the file
/content/<TAXONOMY>/<TERM>/_index.md

Then you will be able to access the taxonomy term as a page in a
taxonomy terms template.

That is why I added instructions in /content-management/taxonomies on
how to add custom metadata. And in /templates/taxonomy-templates/ I
added an example on how you can display that metadata in the taxonomy
terms template.

I think that would be enough to guide the user in how to add custom
metadata to their taxonomy terms.