Add support for --destination sftp://.. style urls the hugo cmd
This commit adds support for using sftp://user:pwd@host/destination/path style urls as
the --destination flag when building a site with hugo.

The sftp implementation is based upon spf13/afero/sftpfs file system. The Hugo command
checks if the --destination (-d) flag provided is a valid sftp url. If so it  initializes
a connection to the server and a afero/sftpfs and sets a hugo fs with the sftpfs as the
Destination.

If username and/or password are not provided within the url the systemuser name will be
used and authentication is tried with the ~/.ssh/id_rsa private key.

Host certificates are evaluated against ~/.ssh/known_hosts