resource: Add preliminary EXIF support
Enable with `imaging.hugo041ExperimentEnableExif = true`.

This commit  also make the Page.Data field into a method.

This is where we put resource-contextual data, such as `Exif` for images.

Having at as an map has some benefits in the templates, as you don't have to know what type of resource it is:

```bash
{{ with .Data.Exif }}Has Exif: {{ .Long }}{{ else }}No Exif for you.{{ end }}
```

The above can be used on any resource type, even pages.

Fixes #4600