Change internal getPage method signature (ref renovations 2 of 6)
Lays the groundwork for ambiguity detection and unix path semantics by
changing the method signature for the core internal page lookups method
used by public API such as the `.GetPage` method and the `ref` and
`relref` shortcodes in advance of the more involved follow on commits.
The change has no functional impact.