helpers/content.go: call rst2html directly to render content, fix #5167
Hugo launches rst2html by calling Python interpreter and passing it the path of rst2html file.
This doesn't work on NixOS like OSs because the real rst2html is actually wrapped behind a bash script which is responsible for correctly setting the env and then exec-ing the real rst2html.

The problem is that on such systems, rst2html is actually a bash script and not a Python script. Hence, Python fails to launch it with SyntaxError.

Ideally, rst2html should be launched directly by Hugo.
This patch ensures that rst2html is called correctly by Hugo.