Fix for hugo issue #4558 related to extracting date and slug from filename for leaf bundles
The definition of the `BaseFilename` field in the `FrontMatterDescriptor` structure provided to the frontmatter handlers has been changed, so that in the case of a leaf bundle `index.`-file, the name provided is that of the containing bundle folder.
Since the use of the `FrontMatterDescriptor` is localized to the fronmatter handler, this should not have external effects.

To compute the new value of `BaseFilename`, I have added a `ContentBaseName` method to the `File` interface.
The `FileInfo` implementation of this interface has been modified to compute the value as part of the `lazyinit`.

Replacing on-demand evaluation of contentBaseName with lazy initialization.

Adding tests