helpers/content.go: call rst2html directly to render content
The current code tries to use the python executable to call
rst2html. This is not correct since the python executable
named `python` might not be the correct python interpreter for the installed
rst2html.
As the code already depend on rst2html to be an executable let's call
rst2html directly.

This fixes as a side-effect also hugo in nixpkgs where rst2html is a shell wrapper.

Co-authored-by: Shreyansh Khajanchi <shreyansh_k@live.com>