Allow date and slug from filename for leaf bundles
To ensure that the front matter handler treats leaf bundle folder names and single file page file
names indentically, the definition of the `BaseFilename` field in the `FrontMatterDescriptor` structure
provided to the frontmatter handlers has been changed, so that in the case of a leaf bundle `index`-file,
the name provided is that of the containing bundle folder.
Since the use of the `FrontMatterDescriptor` is localized to the frontmatter handler, this
should not have external effects.

To compute the new value of `BaseFilename`, a `ContentBaseName` method has been added to the `File` interface.
The `FileInfo` implementation of this interface has been modified to compute the value as part of lazyinit.

Fixes #4558