Update Warning for Theme Links (#676)
The advice in this page is actually the opposite. Relative URLs output the wrong URL for theme assets on the Hugo website.

See relevant advice in the repository's [README](https://github.com/gohugoio/hugoThemes#common-permalink-issues)

I hadn't noticed this page (if I had noticed I would have sent a PR earlier because the advice offered here is plain wrong).

The reason I am merging this PR right away is due to [this](https://github.com/gohugoio/hugoThemes/issues/504#issuecomment-443253163) comment by a confused theme author. 

If you feel that this could be improved feel free to reword and amend as you see fit but the bottom line is that for themes that are going to be published on Hugo Themes authors should create absolute paths in their templates.

cc: @bep @digitalcraftsman @kaushalmodi