Consistently use "posts" as section name in examples
While Hugo does not care if the section is named "post" or "posts",
this commit normalizes the whole Hugo documentation to use "posts"
when used as an example for a section name to prevent user confusion.

Fixes https://github.com/gohugoio/hugoDocs/issues/720.

Also highlight the use of `site.Params.mainSections`.