Stop navigating to draft and future pages on edits with livereload
If hugo was run with
```bash
hugo server --navigateToChanged
```
and a user edited a draft or future page then hugo would navigate to it. Since
drafts don't have urls the navigation would be buggy, sometimes resulting in
404s.
This commit fixes it. Now, the navigation will only be performed only with
--buildDrafts or --buildFuture enabled, like so
```bash
hugo server --navigateToChanged -D
```

Closes #5385