Emit log when omitting paginator output
Before this commit, if the home outputs were configured as:

   home = [ "RSS", "HTML" ]

then a paginator used in the HTML template of the home page would not
generate additional output pages. This is because hugo will only render
the paginator for the main output format of a page, and will silently
ignore the paginators for other output formats.

After this commit, hugo will still ignore those paginators, but will
emit an log message when doing so. This message has log level INFO and
is therefore only visible in verbose mode, to avoid generating false
positives on existing sites where this behaviour is expected.