tpl: Add support to sort and where to access elements of a slice
Template functions sort and where accept a key in order to perform
their taks. However when the key is a slice it is not possible to
access elements using their index. This commit adds support to
access elements of a slice by appending a period followed by the
element's index.

For example:
{{ range (sort .Pages ".Params.dates.0") }}
will sort the pages by using the first element of the slice
.Params.dates