Extend markdownify to support all markup languages
see 5589. We're not introducing another function (htmlify) with a different
name but with almost identical functionality - rather, as we plan to switch to
a generalized .Markup function later on, we'll go with an intermediate
solution. We extend markdownify with an (optional) options map that let's the
user specify the markup format.

See the PR of this commit for a more extensive discussion of the reasoning
behind this.