livereload: Ensure LibreJS compliance
This commit means that when Hugo injects the livereload code onto a page when
the user is using 'hugo server', and a user using LibreJS tries to access a
webpage, the script is not incorrectly flagged as being non-free. It uses special
comments to inform the LibreJS browser extension that Hugo uses the Apache 2.0
license.