Update lookup-order.md
The document starts with "Hugo takes the parameters listed below into consideration when choosing a layout for a given page. They are listed in a priority order."
So it lists parameters from lower priority to higher priority.
But according to examples output format and language priority is higher than Layout priority.
It is quite confusing to read.