Create .editorconfig
Fix for: #6069

I think it'd be a good idea to add an .editorconfig to the projects root for contributors to easily see the code-formatting style. I imagine the file would be something like this but (of course) feel free to make any changes. There is an `.editorconfig` in the HugoDocs repo with the same config rules (https://github.com/gohugoio/hugoDocs/blob/master/.editorconfig)