Add ability to generate nginx 30X configuration
Add the ability to set a new build option `nginxAliases`. If set to true
when building a new `nginx.conf` file will be generated in the project's
root directory. This file will be filled with all configurated aliases
and can be directly include in website nginx configuration.

The main objective of this feature is to prevent the http page loading
(HTTP status of 200) containing only a meta refresh html tag, but
directly serving a 30X HTTP status reponse. This is mainly SEO related.