tpl: Support `start` and `end` parameters in `youtube` shortcode
Adds the `start` and `end` parameters to the `youtube` shortcode so that it's possible to embed shorter snippets of a longer youtube video.

This works both with and without `autoplay="true"`.

See https://github.com/gohugoio/hugo/issues/3694#issuecomment-315802797, which references someone else needing this functionality, and cloning / modifying the existing shortcode for this purpose (though this bug admittedly feels very nebulous!)