Clarify leaf bundle explanation and related FAQ
Organize the content of the leaf bundle such that the example with 4 leaf
bundles shows four headings. Align headings with branch bundles.
Explain that the content files will not result in rendered pages.

Add a hint to check for index.md's that may inadvertently have created
a leaf bundle thus preventing a page from being rendered.