publisher: Don't stop collecting while inside an ignored tag.
This stops parsing of 'preformatted' tags not on any tag found,
but only on finding the matching end tag. Key element here is
looking at the end of a tag (strings.HasSuffix()), rather than
the entire tag. Remainder is simplification of the logic flow.

On what failed before, see also previous commit.

This is related to issue #7567 and fixes the failing test case
introduced with the previous commit.