Remove outdated references in scss-sass.md
It is unclear where the Q1 claim comes from, but it is long past and thus invalidated, hence removed.

Naming the current beta version number adds a requirement to constantly update it, which evidently did not happen. The documentation works just fine without explicitly naming the current beta version number.