sitemap: Hide pages from sitemap if `.Sitemap.Hidden` is true
This commit changes the default `sitemap.xml` to exclude pages from the
sitemap if a page author has set `.Sitemap.Hidden` to `true` in the
front-matter of a page.

The motivation for this change is for cases where a site maintainer
wants to use content from a special category to embed onto another page.
In this case, the content in the special category is already present on
the site, but the pages in the special category are created anyways.
These pages are included in the sitemap and will be indexed by search
engines, but this may not be the desired behavior. So, a site maintainer
is now able to exclude pages from the sitemap without making custom
modifications to the `sitemap.xml` and losing out on any changes made to
the default sitemap in Hugo in the future.

This change was inspired from a support thread from May 2015:

https://discourse.gohugo.io/t/how-to-omit-page-from-sitemap/1126?u=jwf

Documentation is also updated so designers can also take note of this
new change when reading about Hugo's sitemap template.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>